+++ src/components/templates/LayoutTemplate.jsx
@@ -5,9 +5,6 @@ import { useEffect } from "react";
 import HeaderOrganism from "../organism/HeaderOrganism";
 import FooterOrganism from "../organism/FooterOrganism";
 
-// utilities
-import { axiosRes } from "../../utils/axiosDefaults";
-
 // custom hooks
 import useAppContext from "../../hooks/useAppContext";
 import useRefreshToken from "../../hooks/useRefreshToken";
@@ -22,25 +19,6 @@ import styles from "../../styles/templates/Layout.module.css";
  * @constructor
  */
 const LayoutTemplate = ({ children }) => {
-  // state destructuring
-  const { dispatch } = useAppContext();
-
-  useEffect(() => {
-    const handleMount = async () => {
-      const { data } = await axiosRes.get("dj-rest-auth/user/");
-      return data;
-    };
-
-    handleMount()
-      .then((data) => dispatch({ type: "UPDATE USER DATA", payload: data }))
-      .catch((err) => {
-        dispatch({ type: "USER DATA ERROR", payload: "None", error: err });
-      });
-  }, [dispatch]);
-
-  // refreshes the user data tokens
-  useRefreshToken();
-
   return (
     <>
       <HeaderOrganism />
