+++ src/hooks/useRefreshToken.jsx
@@ -1,6 +1,6 @@
 // 3rd party library
 import axios from "axios";
-import { useMemo } from "react";
+import { useEffect, useMemo } from "react";
 
 // utilities
 import { axiosReq, axiosRes } from "../utils/axiosDefaults";
@@ -16,8 +16,26 @@ const useRefreshToken = () => {
   // const history = useNavigate(); // redirect to specific url
 
   // state destructuring
-  const { state } = useAppContext();
+  const { state, dispatch } = useAppContext();
   const { userReducers } = state;
+
+  const handleMount = async () => {
+    return await axiosRes.get("dj-rest-auth/user/");
+  };
+
+  useEffect(() => {
+    handleMount()
+      .then((data) => {
+        dispatch({ type: "UPDATE USER DATA", payload: data });
+        console.log("data", data);
+      })
+      .catch((err) => {
+        dispatch({ type: "USER DATA ERROR", payload: "None", error: err });
+        console.log("error", err);
+      });
+  }, [dispatch]);
+
+  console.log("You are here");
 
   useMemo(() => {
     axiosReq.interceptors.response.use(
@@ -26,13 +44,17 @@ const useRefreshToken = () => {
           await axios.post("/dj-rest-auth/token/refresh/");
         } catch (err) {
           if (userReducers === "None") {
+            console.log("userReducers error");
             // history("/sign-in");
           }
+          console.log("config");
           return config;
         }
+        console.log("config1");
         return config;
       },
       (err) => {
+        console.log("rejected");
         return Promise.reject(err);
       },
     );
@@ -41,17 +63,24 @@ const useRefreshToken = () => {
       (response) => response,
       async (err) => {
         if (err.response?.status === 401) {
-          try {
-            await axios.post("/dj-rest-auth/token/refresh/");
-          } catch (error) {
-            if (userReducers === "None") {
-              // history("/sign-in");
-            }
-            return null;
-          }
+          // this will get rid of the post-unauthorized error in the console
+          console.log("I have made a mistake");
+          // try {
+          //   await axios.post("/dj-rest-auth/token/refresh/");
+          //   console.log("data");
+          // } catch (error) {
+          //   if (userReducers === "None") {
+          //     console.log("error2");
+          //     // history("/sign-in");
+          //   }
+          //   console.log("null");
+          //   return null;
+          // }
 
-          return axios(err.config);
+          // console.log("axios");
+          // return axios(err.config);
         }
+        console.log("rejected2");
         return Promise.reject(err);
       },
     );
